% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{contextify}
\alias{contextify}
\title{contextify}
\usage{
contextify(api, tableName, columnName, table)
}
\arguments{
\item{api}{the instance of the an integer vector of nomisweb geographical codes}

\item{tableName}{name of census table}

\item{columnName}{name of column in the table}

\item{table}{the table}
}
\value{
the table containing a new column with the contextual data
}
\description{
Append table with a contextual column.
}
\examples{
\dontrun{
library("UKCensusAPI")
cacheDir = "/tmp/UKCensusAPI/"
censusapi = UKCensusAPI::instance(cacheDir)
table = "KS401EW"
queryParams = list(
  date = "latest",
  CELL = "7...13", # dwelling type
  select = "GEOGRAPHY_CODE,CELL,OBS_VALUE",
  MEASURES = "20100",
  geography = "1245710558...1245710560",
  RURAL_URBAN = "0"
)
KS401EW = UKCensusAPI::getData(censusapi, table, queryParams)
annotated = contextify(censusapi, table, "CELL", KS401EW)
}
}
