% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{geoCodeLookup}
\alias{geoCodeLookup}
\title{Geographic code lookup}
\usage{
geoCodeLookup(api, geoCodeString)
}
\arguments{
\item{api}{an instance of the API (returned by instance())}

\item{geoCodeString}{the string representation of a geography, e.g MSOA11}
}
\description{
This function returns the nomisweb code for a particular geographic area
}
