% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{getLADCodes}
\alias{getLADCodes}
\title{Map local authority names to nomisweb codes}
\usage{
getLADCodes(api, laNames)
}
\arguments{
\item{api}{an instance of the UKCensusData API.}

\item{laNames}{a string vector of local authority names.}
}
\value{
an integer vector of nomisweb local authority codes
}
\description{
Map local authority names to nomisweb codes
}
\examples{
library(UKCensusAPI)
censusapi = instance("/tmp/UKCensusAPI")
codes = getLADCodes(censusapi, c("Leeds","Bradford"))
}
