% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKCensusAPI.R
\name{instance}
\alias{instance}
\title{get an instance of the python API (required to call any of the functions)}
\usage{
instance(cacheDir, country = "EW")
}
\arguments{
\item{cacheDir}{directory to cache data}

\item{country}{either "EW" (default, nomisweb API), "SC" (NRScotland bulk data), "NI" (NISRA bulk data)}
}
\value{
an instance of one of the python apis
}
\description{
get an instance of the python API (required to call any of the functions)
}
