% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NomiswebApi.R
\name{modifyGeography}
\alias{modifyGeography}
\title{modifyGeography
Takes an existing nomisweb query and modifies the geography parameter according sipplied coverage and resolution}
\usage{
modifyGeography(queryUrl, coverage, resolution)
}
\arguments{
\item{queryUrl}{a predefined query url}

\item{coverage}{a}

\item{resolution}{the nomisweb code for a particular area type (e.g. 297 for MSOA)}
}
\value{
a modified query
}
\description{
modifyGeography
Takes an existing nomisweb query and modifies the geography parameter according sipplied coverage and resolution
}
\examples{
queryUrl = "https://www.nomisweb.co.uk/api/v01/dataset/NM_618_1.data.tsv?CELL=7...13&MEASURES=20100&RURAL_URBAN=0&date=latest&geography=1245710558...1245710660\%2C1245714998...1245714998\%2C1245715007...1245715007\%2C1245715021...1245715022&select=GEOGRAPHY_CODE\%2CCELL\%2COBS_VALUE"
coverage = c("Leeds","Bradford")
resolution = 299
newQueryUrl = modifyGeography(queryUrl, coverage, resolution)
}
