% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKCensusAPI.R
\name{queryInstance}
\alias{queryInstance}
\title{get an instance of the python query (required to call any of the functions)}
\usage{
queryInstance(cacheDir)
}
\arguments{
\item{cacheDir}{directory to cache data}
}
\value{
an instance of the query module
}
\description{
get an instance of the python query (required to call any of the functions)
}
