% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{queryMetadata}
\alias{queryMetadata}
\title{Interactive metadata query}
\usage{
queryMetadata()
}
\description{
This function calls an interactive script where the user selects a table, a geography, and selects fields, optionally filtering by value.
This script will not run in RStudio due to the way it handles standard input. Please run in a standalone R session (or call the python script directly)
}
\examples{
\dontrun{queryMetadata()}
}
