% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NomiswebApi.R
\name{readLADCodes}
\alias{readLADCodes}
\title{Map local authority names to nomisweb codes}
\usage{
readLADCodes(laNames)
}
\arguments{
\item{laNames}{a string vector of local authority names.}
}
\value{
an integer vector of nomisweb local authority codes
}
\description{
Map local authority names to nomisweb codes
}
