% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKCensusAPI.R
\docType{package}
\name{UKCensusAPI-package}
\alias{UKCensusAPI}
\alias{UKCensusAPI-package}
\title{UKCensusAPI}
\description{
R package for automating queries and downloads of UK census data.
}
\details{
How do you reference the other docs?The only function you're likely to need from \pkg{roxygen2} is
\code{\link{roxygenize}}. Otherwise refer to the vignettes to see
how to format the documentation.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
