% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{contextify}
\alias{contextify}
\title{contextify}
\usage{
contextify(api, tableName, columnName, table)
}
\arguments{
\item{api}{the instance of the an integer vector of nomisweb geographical codes}

\item{tableName}{name of census table}

\item{columnName}{name of column in the table}

\item{table}{the table}
}
\value{
the table containing a new column with the contextual data
}
\description{
Append table with a contextual column.
}
