% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{geoCodes}
\alias{geoCodes}
\title{geoCodes
Get nomisweb geographical codes for a region}
\usage{
geoCodes(api, coverage, resolution)
}
\arguments{
\item{api}{the instance of the an integer vector of nomisweb geographical codes}

\item{coverage}{an integer vector of nomisweb geographical codes}

\item{resolution}{the nomisweb code for a particular area type (e.g. 297 for MSOA)}
}
\value{
a compressed string (nomisweb format) containing nomisweb area codes
}
\description{
geoCodes
Get nomisweb geographical codes for a region
}
\examples{
library(UKCensusAPI)
censusapi = instance("/tmp/UKCensusAPI")
coverage = getLADCodes(censusapi, c("City of London"))
resolution = geoCodeLookup(censusapi, "LSOA11")
codes = geoCodes(censusapi, coverage, resolution)
}
