% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{getData}
\alias{getData}
\title{getData()
Fetch and cache census data using a predefined query}
\usage{
getData(api, tableName, query)
}
\arguments{
\item{api}{a predefined query}

\item{tableName}{name of census table (e.g KS401EW)}

\item{query}{query parameters}
}
\value{
a data.frame contraing the downloaded data
}
\description{
Ensure all query numeric parameters are passed as strings (e.g. "0" not 0)
This prevents conversion to floating-point which can makie queries fail
}
\examples{
\dontrun{
library(UKCensusAPI)
censusapi = instance("/tmp/UKCensusAPI")
table = "KS102EW"
meta=getMetadata(censusapi, table)
# queryParams can be autogenerated using the interactive query functionality
queryParams = list(
  geography = "1249902593...1249902596,1249934513...1249934514",
  MEASURES = "20100",
  select = "GEOGRAPHY_CODE,OBS_VALUE",
  CELL = "0",
  RURAL_URBAN = "0",
  date = "latest"
)
getData(censusapi, table, queryParams)
}
}
