% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nomisweb.R
\name{getData}
\alias{getData}
\title{getData()
Fetch and cache census data using a predefined query}
\usage{
getData(api, tableName, internalName, query)
}
\arguments{
\item{api}{a predefined query}

\item{tableName}{name of census table (e.g KS401EW)}

\item{internalName}{internal name of census table (e.g NM_618_1)}

\item{query}{query parameters}
}
\description{
Ensure all query numeric parameters are passed as strings (e.g. "0" not 0)
This prevents conversion to floating-point which can makie queries fail
}
