% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKCensusAPI.R
\name{instance}
\alias{instance}
\title{get an instance of the python API (required to call any of the functions)}
\usage{
instance(cacheDir)
}
\arguments{
\item{cacheDir}{directory to cache data}
}
\value{
an instance of the ukcensusweb api
}
\description{
get an instance of the python API (required to call any of the functions)
}
