% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NomiswebApi.R
\name{queryInstance}
\alias{queryInstance}
\title{get an instance of the python query (required to call any of the functions)}
\usage{
queryInstance(api)
}
\arguments{
\item{api}{an instance of the ukcensusapi}
}
\value{
an instance of the query module
}
\description{
get an instance of the python query (required to call any of the functions)
}
