% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKCensusAPI.R
\docType{package}
\name{UKCensusAPI}
\alias{UKCensusAPI}
\alias{UKCensusAPI-package}
\title{UKCensusAPI}
\description{
R package for automating queries and downloads of UK census data.
}
\details{
See README.md for detailed information and examples.
}
\section{Functions}{

\code{\link{geoCodes}}

\code{\link{getData}}

\code{\link{getLADCodes}}

\code{\link{getMetadata}}

\code{\link{instance}}

\code{\link{queryInstance}}

\code{\link{queryMetadata}}
}

